+++ lib/clients/httpClient_test.py
@@ -1,7 +1,14 @@
 from .httpClient import HttpClient
 import re
 import pytest
-httpClient = HttpClient()
+httpClient = None
+
+
+@pytest.fixture(autouse=True)
+async def run_around_tests():
+    global httpClient
+    httpClient = HttpClient()
+    yield
 
 
 class TestHttpClient:
@@ -26,3 +33,16 @@ class TestHttpClient:
             raise Exception('NotFoundException is expected')
         except Exception as err:
             assert err.__class__.__name__ == 'NotFoundException'
+
+    @pytest.mark.asyncio
+    async def test_timeout(self):
+        """Should return ConnectTimeout exception if request is timed out"""
+        httpClient = HttpClient(0.001)
+        opts = {
+            'url': 'http://example.com/'
+        }
+        try:
+            await httpClient.request(opts)
+            raise Exception('ConnectTimeout is expected')
+        except Exception as err:
+            assert err.__class__.__name__ == 'ConnectTimeout'